+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Nix package to flake.nix.
+- Command-line and configuration file option for setting the initial tab.
 
 ### Fixed