+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Combine bindings for opening a dropdown and choosing a dropdown item.
+
 ## [0.3.0] - 2025-05-13
 
 ### Added